remove dir

+++ utils/data_loader.py
@@ -170,7 +170,8 @@ class DataLoader:
                 pickle.dump((train_data, eval_data, vocab), f)
             tf.logging.info('Refreshed data successfully dumped at {}'.format(FLAGS.prc_data_loc))
 
-            shutil.rmtree(FLAGS.output_dir)
+            if os.path.exists(FLAGS.output_dir):
+                shutil.rmtree(FLAGS.output_dir)
             os.mkdir(FLAGS.output_dir)
         else:
             tf.logging.info('Restoring data from {}'.format(FLAGS.prc_data_loc))
